feat(home): #SKFP-5 refactor login module
