+++ src/App.js
@@ -20,9 +20,9 @@ import ParticipantEntity from 'components/EntityPage/Participant';
 import Error from 'components/Error';
 import FenceAuthRedirect from 'components/Fence/FenceAuthRedirect';
 import FileRepo from 'components/FileRepo';
-import Join from 'components/Login/Join';
-import LoginPage from 'components/Login/LoginPage';
-import TermsConditions from 'components/Login/TermsConditions';
+import LoginPage from 'components/LoginPage';
+import Join from 'components/LoginPage/Join';
+import TermsConditions from 'components/LoginPage/TermsConditions';
 import MemberSearchPage from 'components/MemberSearchPage';
 import Page, { FixedFooterPage } from 'components/Page';
 import SideImagePage from 'components/SideImagePage';
