--- src/components/Login/index.css
@@ -1,23 +1,0 @@
-.login-redirect-loader-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-}
-
-.login-redirect-loader-container > span {
-  color: #2b388f;
-  font-size: 1.1em;
-  font-weight: 500;
-  margin-top: 1.25em;
-}
-
-.login-redirect-loader-container .generic-error-display a {
-  color: #a6278f;
-}
-.login-redirect-loader-container .generic-error-display a:hover {
-  color: #cc3399;
-}
-
-.text-align-center {
-  text-align: center;
-}
