+++ src/components/LoginPage/Login/LoginStats/index.module.scss
@@ -1,0 +1,44 @@
+@import 'src/style/themes/default/colors';
+
+.loginStats {
+  background-color: $cyan-8;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: flex-start;
+  padding: 24px;
+  margin: 48px 0px;
+
+  hr {
+    border: none;
+    border-top: 1px solid $cyan-7;
+    width: 100%;
+  }
+
+  h4 {
+    color: $cyan-4;
+    font-style: normal;
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 24px;
+    margin: 8px 0px;
+  }
+
+  > div {
+    width: 100%;
+  }
+
+  .loginStatsContainer {
+    justify-content: space-between;
+    display: flex;
+    margin-top: 24px;
+
+    .loginPageIconColor {
+      fill: $cyan-5;
+    }
+
+    .loginStatsLabel {
+      color: $gray-1;
+    }
+  }
+}
