+++ src/components/LoginPage/LoginFooter/index.module.scss
@@ -1,0 +1,25 @@
+@import 'src/style/themes/default/colors';
+
+.loginFooter {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  background-color: $gray-1;
+  padding: 24px;
+
+  h3 {
+    margin-bottom: 16px;
+    font-weight: 600;
+    font-size: 16px;
+    line-height: 24px;
+    color: $body-1;
+  }
+
+  span {
+    margin-bottom: 40px;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 22px;
+    color: $body-1;
+  }
+}
