+++ src/components/LoginPage/TermsConditions/index.tsx
@@ -2,15 +2,13 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { Button } from 'antd';
 
+import TermsConditionsAcceptButton from 'components/LoginPage/TermsConditions//TermsConditionsAcceptButton';
+import TermsConditionsBody from 'components/LoginPage/TermsConditions//TermsConditionsBody';
+import SplashPage from 'components/SplashPage';
 import { revertAcceptedTermsThenLogoutCleanly } from 'store/actionCreators/user';
 import { DispatchUser } from 'store/userTypes';
 
-import SplashPage from '../SplashPage';
-
-import TermsConditionsAcceptButton from './TermsConditionsAcceptButton';
-import TermsConditionsBody from './TermsConditionsBody';
-
-import './termsConditions.css';
+import './index.css';
 
 const TermsConditions = () => {
   const dispatch = useDispatch<DispatchUser>();
