+++ src/components/LoginPage/index.module.scss
@@ -1,0 +1,8 @@
+.loginPage {
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  width: 100%;
+  justify-content: space-between;
+  padding: 24px;
+}
