+++ src/components/LoginPage/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import Login from 'components/Login/Login';
-import LoginFooter from 'components/Login/LoginFooter';
+import Login from 'components/LoginPage/Login';
+import LoginFooter from 'components/LoginPage/LoginFooter';
 
-import styles from './Login.module.scss';
+import styles from './index.module.scss';
 
 const LoginPage = () => (
   <div className={styles.loginPage}>
